Reorganize security modules into logical namespaces for better maintainability
This comprehensive reorganization restructures Otto's security module for improved
maintainability, separation of concerns, and future scalability:

## New Directory Structure:
- lib/otto/security/authentication/ - All authentication components
  - auth_strategy.rb - Base strategy class
  - authentication_middleware.rb - Auth enforcement middleware
  - strategy_result.rb & failure_result.rb - Result objects
  - strategies/ - Concrete strategy implementations
- lib/otto/security/middleware/ - All security middleware
- Index files provide backward compatibility aliases

## Key Changes:
- Moved StrategyResult from Otto:: to Otto::Security::Authentication::
- Moved all strategies to Otto::Security::Authentication::Strategies::
- Moved all middleware to Otto::Security::Middleware::
- Added comprehensive backward compatibility through aliases
- Updated all internal references to use new namespaces
- Extracted authentication middleware to separate file
- Split rate limiting logic from middleware

## Benefits:
- Clear separation of authentication, middleware, and utilities
- Scalable structure for adding new strategies/middleware
- Maintains 100% backward compatibility
- Follows Ruby conventions and single responsibility principle
- Easier to locate and maintain specific components
