+++ lib/otto/security/authentication/strategies/public_strategy.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require_relative '../auth_strategy'
+require_relative '../strategy_result'
+
+class Otto
+  module Security
+    module Authentication
+      module Strategies
+        # Public access strategy - always allows access
+        class PublicStrategy < AuthStrategy
+          def authenticate(env, _requirement)
+            Otto::Security::Authentication::StrategyResult.anonymous(metadata: { ip: env['REMOTE_ADDR'] })
+          end
+        end
+      end
+    end
+  end
+end
